requirements.txt
Flask==2.1.1
flask-cors==3.0.10
python-dotenv==0.20.0
pymongo[srv]==4.1.1
Werkzeug==2.0.3
pyjwt==2.4.0
flask-limiter==2.1.0
bleach==5.1.0